pyproject.toml
[tool.poetry]
name = "bookit"
version = "0.1.0"
description = "A space booking management system written in Django"
authors = ["Dylan Wilson <dylanjw@protonmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
Django = "=3.0"
pip = "^19.3.1"
django-environ = "^0.4.5"
django-reservation = "^0.2.9"
django-scheduler = "^0.9.3"
django-yarn = "^0.0.2"
django-webpack-loader = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.1"
iPython = "^7.10"
flake8 = "^3.7.9"
pytest-django = "^3.7.0"
pytest-pythonpath = "^0.7.3"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
